1.4.0
Added the Advanced Melee Attack Macro by Jules to the D&D5e Animations Macros compendium so it can be used to add trails and other effects to weapons that use the new animations. Some example items have been added to a second compendium, look at the A-A settings of the items for the macro arguments.

New Animations

-Melee
--Baton
--Bone
--Butterfly
--Cutlass
--Dagger
--Katana
--Katar
--Knife
--Machete
--Naginata
--Ninjato
--Odachi (Nodachi also triggers it)
--Shield (excludes spells)
--Strike
--Tanto

-Ranged
--Antipathy/Sympathy
--Boomerang
--Kunai
--Missile
--Planar Binding
--Shuriken
--Snow Ball

-On Token
--Animate Objects
--Arcane/Bigby's Hand
--Arcane/Mordenkainen's Sword
--Awaken
--Clone
--Demiplane
--Faithful Hound
--Forbiddance
--Glibness
--Guards and Wards
--Magic Jar
--Magnificent Mansion
--Mass Heal
--Secret Chest
--Sequester
--Shapechange
--Simulacrum
--Transport via Plants
--Word of Recall

-Template
--Fabricate
--Forcecage
--Mass Cure Wounds
--Private Sanctum
--Reverse Gravity
--Spiritual Weapon (for SRD version)
--Storm of Vengeance
--Weird

Changed Animations

-Melee

--Scythe (animation)
--Sickle (animation)

-Ranged
--Bow (sound)
--Crossbow (sound)
--Gun (sound)
--Musket (sound)
--Revolver (sound)
--Rifle (sound)
--Pistol (sound)

-On Token
--Flame Blade (set to exact match)
--Green-Flame Blade (set to exact match)
--Heal (set to exact match)
--Spiritual Weapon (overhauled)

-Active Effects
--Shield (now excludes equipment)

Removed Animations

-On Token
--Shield (reduntant with Active Effect)

-Active Effects
--Charmed
--Deafended
--Frightened
--Grappled
--Incapacitated
--Paralyzed
--Poisoned
--Prone
--Restrained
--Stunned
--Unconscious

Most of the removals were status conditions that were added earlier that I'll admit were getting on my nerves and caused other problems so I wanted to remove them. You can add them yourself if you wish and they won't be overwritten, but they will no longer be in the module.